Reorganize express app initialization
